make more compatible to custom C128D setup
This change is for a custom C128D setup with the SD2IEC being installed
internally (like a hard drive) with no buttons to the outside. A long
press on restore triggers the "next image" button, and the "prev image"
button is connected to the IEC reset line. This leads to the following
functional changes of the code.

- SD2IEC can be taken off the IEC bus by sending the command "H"
- detecting IEC reset does the "unhide" and also re-reads the config
  mainly to reset the drive id, so a reset works similar to the reset
  of the internal 1571 drive
- this behavior is configured by setting CONFIG_USE_PREV_FOR_RESET